Move `runCLI` into the cli folder.
