+++ packages/jest-cli/src/jest.js
@@ -9,94 +9,13 @@
  */
 'use strict';
 
-import type {AggregatedResult} from 'types/TestResult';
-import type {Path} from 'types/Config';
-
-const realFs = require('fs');
-const fs = require('graceful-fs');
-fs.gracefulify(realFs);
-
-const Runtime = require('jest-runtime');
 const SearchSource = require('./SearchSource');
 const TestRunner = require('./TestRunner');
-
-const chalk = require('chalk');
-const {Console, clearLine} = require('jest-util');
-const {createDirectory} = require('jest-util');
-const createHasteContext = require('./lib/createHasteContext');
-const getMaxWorkers = require('./lib/getMaxWorkers');
-const logDebugMessages = require('./lib/logDebugMessages');
-const preRunMessage = require('./preRunMessage');
-const readConfig = require('jest-config').readConfig;
-const {run} = require('./cli');
-const runJest = require('./runJest');
 const TestWatcher = require('./TestWatcher');
-const watch = require('./watch');
+
+const {run, runCLI} = require('./cli');
 
 const VERSION = require('../package.json').version;
-
-const runCLI = (
-  argv: Object,
-  root: Path,
-  onComplete: (results: ?AggregatedResult) => void,
-) => {
-  const pipe = argv.json ? process.stderr : process.stdout;
-  argv = argv || {};
-  if (argv.version) {
-    pipe.write(`v${VERSION}\n`);
-    onComplete && onComplete();
-    return;
-  }
-
-  const _run = async ({config, hasDeprecationWarnings}) => {
-    if (argv.debug) {
-      logDebugMessages(config, pipe);
-    }
-
-    createDirectory(config.cacheDirectory);
-    const hasteMapInstance = Runtime.createHasteMap(config, {
-      console: new Console(pipe, pipe),
-      maxWorkers: getMaxWorkers(argv),
-      resetCache: !config.cache,
-      watch: config.watch,
-    });
-
-    const hasteMap = await hasteMapInstance.build();
-    const hasteContext = createHasteContext(config, hasteMap);
-    if (argv.watch || argv.watchAll) {
-      return watch(
-        config,
-        pipe,
-        argv,
-        hasteMapInstance,
-        hasteContext,
-        hasDeprecationWarnings,
-      );
-    } else {
-      const startRun = () => {
-        preRunMessage.print(pipe);
-        const testWatcher = new TestWatcher({isWatchMode: false});
-        return runJest(
-          hasteContext,
-          config,
-          argv,
-          pipe,
-          testWatcher,
-          startRun,
-          onComplete,
-        );
-      };
-      return startRun();
-    }
-  };
-
-  readConfig(argv, root).then(_run).catch(error => {
-    clearLine(process.stderr);
-    clearLine(process.stdout);
-    console.error(chalk.red(error.stack));
-    process.exit(1);
-  });
-};
 
 module.exports = {
   SearchSource,
